--- src/logic/version/VERSION.txt
@@ -1,1 +1,0 @@
-299843
