--- src/logic/version/updater_VERSION.txt
@@ -1,1 +1,0 @@
-1166
